@@ -1,15 +1,10 @@
 import 'normalize.css'
 import 'remconvert'
 import './index.css'
-import App from './App';
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
+import App from './App';
 
-
-ReactDOM.render(
-  <>
-    <App />
-  </>,
-  document.getElementById('root')
-);
+const root = ReactDOM.createRoot(document.getElementById('root'))
+root.render(<App/>)
 
